==> 2 thomson/2_thomson_get_freq.ipynb <==
# Firm 별 투자 Freq 계산

## 1. vehicle patent에 해당하는 row 찾기  -> 이제 보니 쓸모없음 
우리 scope대로 vehicle 관련 company에 투자한 정보만 찾아야지

import pandas as pd
import zipfile
import csv
import os

os.chdir("E:/patent")
zf_patent1 = zipfile.ZipFile('all_patent/harvard_dataset_final_merged_output_1.zip') 
csv_patent1 = zf_patent1.open('harvard_dataset_final_merged_output_1.csv')
csv_patent_reader1 = csv.reader(csv_patent1)

zf_patent2 = zipfile.ZipFile('all_patent/harvard_dataset_final_merged_output_2.zip')
csv_patent2 = zf_patent2.open('harvard_dataset_final_merged_output_2.csv')
csv_patent_reader2 = csv.reader(csv_patent2)

header = csv_patent_reader1.next()
csv_patent_reader2.next()
mainclass_col = []
for j in range(62457,63354):
    if 'mainclass' in header[j]:   
        mainclass_col.append(j)

classinfo = pd.read_csv('veh_class.csv')
veh_class = classinfo['class'].tolist()
veh_class = map(int,veh_class)

n = 0
row = []
for line in csv_patent_reader1:
    for i in mainclass_col:
        if line[i]=='':
                break
                
        else:
            try:            
                if int(line[i]) in veh_class:
                    row.append(n)
                    break
                
            except ValueError:
                pass    
    n = n+1
    
for line in csv_patent_reader2:
    for i in mainclass_col:
        if line[i]=='':
                break
                
        else:
            try:            
                if int(line[i]) in veh_class:
                    row.append(n)
                    break
                
            except ValueError:
                pass    
    n = n+1

rows=zip(row)
with open('veh_row.csv', 'wb') as f:
    writer = csv.writer(f)
    for val in rows:
        writer.writerow(val)

len(row)

# 2. row에 맞는 투자 정보 추출

#basic settings
import pandas as pd
import zipfile
import csv
import os

os.chdir("E:/patent")

zf_patent1 = zipfile.ZipFile('all_patent/thomson_harvard_matched.zip') 
csv_patent1 = zf_patent1.open('thomson_harvard_matched.csv')
csv_patent_reader1 = csv.reader(csv_patent1)
header = csv_patent_reader1.next()

firm_col = []
for i in range(0,len(header)):
    if 'Firm Name' in header[i]:   
        firm_col.append(i)

row = pd.read_csv('veh_pid_assignee.csv')
veh_row = row.iloc[:,2].tolist()
veh_row = map(int,veh_row)
veh_row = sorted(list(set(veh_row)))

firm = []
assignee = []
amt = []
date = []
n=0

for line in csv_patent_reader1:
    if n == veh_row[0]:        
        for i in firm_col:
            if line[i]=='':
                break

            else:
                firm.append(line[i])
                assignee.append(line[i-24])
                amt.append(line[i+15])
                date.append(line[i+33])
        del veh_row[0]
        if len(veh_row) == 0:
            break
        else:
            n=n+1
    else:
        n=n+1
        
df = pd.DataFrame({'firm':firm,'assignee':assignee,'amount':amt,'date':date})
df = df.drop_duplicates()

amount의 missing value, 변수 type 처리 필요  
('-' -> NaN 처리, str -> float 처리)  

import numpy as np

for i in range(0,len(df.index)):
    if df.iloc[i,0] == '-':
        df.iloc[i,0] = np.nan
    else:
        df.iloc[i,0] = float(df.iloc[i,0])

for i in range(74000,len(df.index)):
    if df.iloc[i,0] == '-' or df.iloc[i,0] == '':
        df.iloc[i,0] = np.nan
    else:
        df.iloc[i,0] = float(df.iloc[i,0])

df.to_csv('veh_firm_assignee_amt_date.csv',index=False)

amt_col=df.groupby('firm')['amount'].sum()
cnt_col=df.groupby('firm').size()
freq=pd.concat([amt_col, cnt_col], axis=1).reset_index()
freq.columns = ['firm', 'amount','count']
freq.to_csv('veh_investment_firm_freq.csv',index=False)

### 그냥 veh_invest 정보 다 담은 csv 만들기

#basic settings
import pandas as pd
import zipfile
import csv
import os

os.chdir("E:/patent")

zf_patent1 = zipfile.ZipFile('all_patent/thomson_harvard_matched.zip') 
csv_patent1 = zf_patent1.open('thomson_harvard_matched.csv')
csv_patent_reader1 = csv.reader(csv_patent1)
header = csv_patent_reader1.next()

row = pd.read_csv('veh_pid_assignee.csv')
veh_row = row.iloc[:,2].tolist()
veh_row = map(int,veh_row)
veh_row = sorted(list(set(veh_row)))

n=0

with open('veh_investment.csv','wb') as myfile:
    myfile_writer = csv.writer(myfile,quoting=csv.QUOTE_ALL)
    myfile_writer.writerow(header)
    for line in csv_patent_reader1:
        if n == veh_row[0]:
            if line[36]!='':
                myfile_writer.writerow(line)
                    
            del veh_row[0]
            if len(veh_row) == 0:
                break
            else:
                n=n+1
        else:
            n=n+1 

잠깐 interrupt 후 다시 시작  
(n=308319: 0~308318번째 row까지는 확인 완료, 308319번째 row부터 ㄱㄱ

for i in range(0,308319):
    csv_patent_reader1.next()

del veh_row[0:68832]
n=308319

with open('veh_investment.csv','ab') as myfile:
    myfile_writer = csv.writer(myfile,quoting=csv.QUOTE_ALL)
    myfile_writer.writerow(header)
    for line in csv_patent_reader1:
        if n == veh_row[0]:
            if line[36]!='':
                myfile_writer.writerow(line)
                    
            del veh_row[0]
            if len(veh_row) == 0:
                break
            else:
                n=n+1
        else:
            n=n+1
