==> 2 thomson/3_get selected investment deal info.ipynb <==
# raw data에서 Investment 변수 생성에 필요한 info만 걸러내기

### raw data(investment)에서 필요한 정보

1. amount(이미 완료)
2. investor type
3. investment date (이미 완료) 
4. investor country
5. investor state (US 한정)
6. company(assignee) founded year
7. company(assignee) age
8. company(assignee) country
9. company state (US 한정)

import pandas as pd
import zipfile
import csv
import os

os.chdir("E:/patent")

zf_patent1 = zipfile.ZipFile('veh_investment.zip') 
csv_patent1 = zf_patent1.open('veh_investment.csv')
csv_patent_reader1 = csv.reader(csv_patent1)
header = csv_patent_reader1.next()

for i in range(0,97):
    print i, header[i]

firm_col = []
for i in range(0,len(header)):
    if 'Firm Name' in header[i]:   
        firm_col.append(i)

firm = []
assignee = []
amt = []
date = []
inv_type = []
fnd_type = []
inv_country = []
inv_state = []
ass_fyear = []
ass_age = []
ass_country = []
ass_state = []

for line in csv_patent_reader1:
    for i in firm_col:
        if line[i]=='':
            break

        else:
            firm.append(line[i])
            assignee.append(line[i-24])
            amt.append(line[i+15])
            date.append(line[i+33])
            inv_type.append(line[i+8])
            fnd_type.append(line[i+30])
            inv_country.append(line[i+1])
            inv_state.append(line[i+4])
            ass_fyear.append(line[i-31])
            ass_age.append(line[i-35])
            ass_country.append(line[i-23])
            ass_state.append(line[i-21])

df = pd.DataFrame({'firm':firm,'assignee':assignee,'amount':amt,'date':date,'firm_type':inv_type,'firm_country':inv_country,
                   'firm_state':inv_state,'assignee_founded_year':ass_fyear,'assignee_age_month':ass_age,'assignee_country':ass_country,
                  'assignee_state':ass_state,'fund_type':fnd_type})

df.to_csv('veh_selected_deal_info_duplicates.csv',index=False)

cols=list(df)
del cols[-1]
cols

df_nodups = df.drop_duplicates(subset=cols)

len(df_nodups.index)

df_nodups.to_csv('veh_sel_info_temp.csv',index=False)

missing value, 변수 type 처리 필요   
('-' -> NaN 처리, str -> float 처리)

def bye_missing_numeric(df,col):
    for i in range(0,len(df.index)):
        for j in col:
            if df.iloc[i,j] == '-' or df.iloc[i,j] == '':
                df.iloc[i,j] = np.nan
                
def bye_missing_string(df,col):
    for i in range(0,len(df.index)):
        for j in col:
            if df.iloc[i,j] == '-' or df.iloc[i,j] == '':
                df.iloc[i,j] = ''

import numpy as np

col = range(0,11)
num_col = [0,2,4]
str_col = [x for x in col if x not in num_col]
bye_missing_numeric(df,num_col)
bye_missing_string(df,str_col)

df.iloc[0:5,]

df.iloc[:,0]=map(float,df.iloc[:,0])

for i in range(0,len(df.index)):
    for j in [2,4]:
        if pd.isnull(df.iloc[i,j])==False:
            df.iloc[i,j]=int(df.iloc[i,j])

print type(df.iloc[0,0]), type(df.iloc[0,2]), type(df.iloc[2,2]), type(df.iloc[2,4])

def nameprocessing(s):
    s = str(s)
    
    s = s.replace(".","")
    s = s.replace(",","")    
    s = s.replace(";","")
    s = ''.join(s.split()).lower()
    
    s = s.replace("incorporated","inc")
    s = s.replace("limited","ltd")
    s = s.replace("gmbh","")
    
    rm=['inc','ltd','llc','pte']
    for i in range(0,len(s)):
        if s[-3:] in rm:
            s=s[:-3]
        else:
            break

    return s

df['new_assignee']=''
df.iloc[0:5,]

for i in range(0,len(df.index)):
    df.iloc[i,11]=nameprocessing(df.iloc[i,1])

df.to_csv('veh_selected_deal_info.csv',index=False)
