==> 2 thomson/4_Investment variable generation.ipynb <==
# Invest info로 파생 변수 생성


### company, year 기준으로 group by
1. amount: sum
2. investor type: dummy, count
3. investment year: date -> year
4. investor country : dummy, count
5. investor state (US 한정) : dummy, count
6. company(assignee) age (year):investment year - founded year
7. company(assignee) country : dummy, count
8. company state (US 한정) : dummy, count
9. investor-company distance(US 한정) - __새로 추가__

import pandas as pd
import os
import numpy as np

os.chdir("E:/patent")

df = pd.read_csv('veh_selected_deal_info_final.csv')

## 1. Dummy들 만들기 
1. investor type
2. investor country
3. investor state

### 1.1 investor type
1. 어떤 type이 있나?  
2. 각 type별 투자 횟수
  
__구글 refine 쓰면 편할듯?__

__Firm type__  

Private Equity Firm 61283  
Other 7498  
Corporate PE/Venture 7286  
Bank Affiliated 5238  
Individuals 2511  
Investment Management Firm 1370  
Government Affiliated Program 853  
Endowment, Foundation or Pension Fund 628  
Private Equity Advisor or Fund of Funds 628  
SBIC 421  
Insurance Firm Affiliate 378  
Angel Group 254  
Incubator/Development Program 246  
University Program 223  
Service Provider 131  
Non-Private Equity 59   
  
__Fund type__  
Venture Capital 76823  
Generalist Private Equity 5286  
Buyout 5118  
Fund of Funds 838  
Mezzanine 612  
Other Private Equity 249  
Non-Private Equity 59  
includeeditReal Estate 21  

Private Equity & Venture Capital: 53227

### 1.2 investor country  
1. 어떤 나라가 있나?
2. 나라별 투자 횟수

United States 69193  
Unknown 6697  
Canada 3348  
United Kingdom 2608  
Germany 1058  
Israel 765  
Japan 538  
France 537  
South Korea 448  
Switzerland 421  
Taiwan 407  
Sweden 394  
Singapore 297  
Australia 276  
Netherlands 214  
China 201  
Denmark 195  
Norway 190  
Finland 177  
Hong Kong 159  
Belgium 146  
Italy 96  
Ireland 87  
Luxembourg 71  
India 69  
Russia 61  
Spain 51  
Austria 41  
Bahrain 27  
Cayman Islands 27  
Bermuda 23  
Mauritius 19  
Brazil 16  
Utd. Arab Em.16  
Malaysia 15  
New Zealand 15  
Philippines 10  
Portugal 10   
Czech Republic9
Channel Islands8
Saudi Arabia 7
Argentina6
Kuwait6
Poland6
South Africa6
Greece5
Iceland3
Liechtenstein3
Mexico3
Bulgaria2
Cyprus2
Estonia2
Br. Virgin I.1
Colombia1
Egypt1
Georgia1
Hungary1
Indonesia1
Latvia1
Lebanon1
Peru1
Qatar1
Thailand1
Tunisia1
Turkey1
Uruguay1
Vietnam1
Zimbabwe1

### 1.3 investor state  
assignee, investor 모두 구글 refine으로 non-US = null 처리   
Alaska, Mississippi, North Dakota = 0

df.iloc[0:3,]

dummy = pd.get_dummies(df, dummy_na=True, columns=['firm_country','firm_state','firm_type','fund_type'])
dummy.iloc[1,]

dummy.to_csv('veh_selected_deal_info_dummies.csv',index=False)

## 2. 파생변수 만들기

1. investment year: date -> year
2. company(assignee) age (year):investment year - founded year
3. investor-company distance

### 2.1 investment year
구글 refine으로 처리

### 2.2 assignee age(year)

list(df)

df['assignee_age_year']=df['invest_year']-df['assignee_founded_year']
df.iloc[14:23,]

cols=list(df)
cols

newcols = cols[0:2]+cols[-1:]+cols[2:-1]
df = df[newcols]
df.iloc[0:3,]

df.to_csv('veh_selected_deal_info.csv',index=False)

### 2.3 investor-company distance

import pandas as pd
import os
import geopy as gp
import numpy as np

os.chdir("E:/patent")
df = pd.read_csv('veh_selected_deal_info_final.csv')
caps = pd.read_csv('US_state_capital.csv')

list(df), list(caps)

ndf = pd.merge(df,caps,how='left',left_on='assignee_state', right_on='state', sort=False)

ndf.iloc[0:4,]

del ndf['state']

ndf.rename(columns={'capital':'assingee_capital','latitude':'assingee_latitude','longitude':'assingee_longitude'}, inplace=True)
list(ndf)[-3:]

ndf2 = pd.merge(ndf,caps,how='left',left_on='firm_state', right_on='state', sort=False)

ndf2.iloc[0:4,]

del ndf2['state']
ndf2.rename(columns={'capital':'firm_capital','latitude':'firm_latitude','longitude':'firm_longitude'}, inplace=True)
list(ndf2)[-3:]

ndf2.sort_index(axis=1, inplace=True)
list(ndf2)

df=ndf2
del ndf, ndf2

df['assignee_firm_dist']=np.nan

df.iloc[0:2,]

from geopy.distance import great_circle

a_loc = (df['assingee_latitude'][0],df['assingee_longitude'][0])
f_loc = (df['firm_latitude'][0],df['firm_longitude'][0])
great_circle(a_loc,f_loc).kilometers

df.loc[0,'assignee_firm_dist'] = great_circle(a_loc,f_loc).kilometers
df.iloc[0,]

for i in range(0,len(df)):
    if pd.isnull(df['firm_latitude'][i]) == False:
        a_loc = (df['assingee_latitude'][i],df['assingee_longitude'][i])
        f_loc = (df['firm_latitude'][i],df['firm_longitude'][i])
        df.loc[i,'assignee_firm_dist'] = great_circle(a_loc,f_loc).kilometers

df.iloc[(0,3,90,108,273),]

df.to_csv('veh_selected_deal_info_final2.csv',index=False)

# 3. groupby assignee, year

### groupby 시 사용할 함수
1. sum: amt, dummies  
2. mean: age, distance  
3. first: ass_country, ass_founded_year, ass_state

__distance는 sum? mean?__ => mean!

import pandas as pd
import os
import numpy as np

os.chdir("E:/patent/thomson")
df = pd.read_csv('veh_selected_deal_info_dummies.csv')

### 3.1 함수에 따라 변수 분리 (dummy들 때문에 한 번에 돌리기 힘듦)

col = []
for i in range (0,len(list(df))):
    col.append(i)
max(col)

meancol = [1,2,3,7,9]
firstcol = [1,4,5,6,9]
trashcol = [0,8,11]
sumcol = [1,9]+list(set(col)-set(meancol)-set(firstcol)-set(trashcol))

sumdf = df.iloc[:,sumcol]
meandf = df.iloc[:,meancol]
firstdf = df.iloc[:,firstcol]

list(sumdf)[0:2]

sums = sumdf.groupby(['new_assignee','invest_year'],as_index=False).sum()
means = meandf.groupby(['new_assignee','invest_year'],as_index=False).mean()
firsts = firstdf.groupby(['new_assignee','invest_year'],as_index=False).first()

print len(sums), len(means), len(firsts)

df = firsts.merge(means,how='left',on=['new_assignee','invest_year'],sort=True).merge(
    sums,how='left',on=['new_assignee','invest_year'],sort=True)

len(df)

df.to_csv('veh_investment_groupby_assignee_year.csv',index=False)
